feat: Restore error component
