+++ src/pages/UserProfilePage.jsx
@@ -25,7 +25,7 @@ import { postUserPolicy, cullOldPolicies } from "../api/userPolicies";
 import { COUNTRY_NAMES } from "../data/countries";
 import moment from "moment";
 import { formatCurrencyAbbr } from "../lang/format";
-import ErrorPage from "../layout/Error";
+import ErrorPage from "../layout/ErrorPage";
 
 const STATES = {
   EMPTY: "empty",
